backend/requirements.txt
garminconnect>=0.1.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
pandas>=1.3.0
python-dotenv>=0.19.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
typer>=0.9.0